+++ indra-mongo/src/main/java/org/lambda3/indra/mongo/MongoIndraDriver.java
@@ -26,88 +26,17 @@ package org.lambda3.indra.mongo;
  * ==========================License-End===============================
  */
 
-import org.apache.commons.math3.linear.ArrayRealVector;
-import org.apache.commons.math3.linear.OpenMapRealVector;
-import org.apache.commons.math3.linear.RealVector;
-import org.lambda3.indra.client.MutableAnalyzedTerm;
-import org.lambda3.indra.client.ScoreFunction;
-import org.lambda3.indra.client.TextPair;
-import org.lambda3.indra.core.*;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import org.lambda3.indra.core.IndraDriver;
+import org.lambda3.indra.core.Params;
 
-import java.io.IOException;
-import java.util.HashMap;
-import java.util.LinkedList;
-import java.util.List;
-import java.util.Map;
-
-public class MongoIndraDriver {
-
-    private static final String DEFAULT_CORPUS_NAME = "wiki-2014";
-    private static final String DEFAULT_LANGUAGE = "EN";
-    private static final String DEFAULT_DISTRIBUTIONAL_MODEL = "W2V";
-    private static final Params DEFAULT_PARAMS = new Params(DEFAULT_CORPUS_NAME, ScoreFunction.COSINE,
-            DEFAULT_LANGUAGE, DEFAULT_DISTRIBUTIONAL_MODEL);
-
-    private Logger logger = LoggerFactory.getLogger(getClass());
-
-    private VectorSpaceFactory vectorSpaceFactory;
-    private RelatednessClientFactory relatednessClientFactory;
-    private Params currentParams;
+public class MongoIndraDriver extends IndraDriver {
 
     public MongoIndraDriver(String mongoURI) {
-        this(DEFAULT_PARAMS, new MongoVectorSpaceFactory(mongoURI));
+        this(IndraDriver.DEFAULT_PARAMS, mongoURI);
     }
 
-    public MongoIndraDriver(Params params, VectorSpaceFactory vectorSpaceFactory) {
-        this.currentParams = params;
-        this.vectorSpaceFactory = vectorSpaceFactory;
-        //this.relatednessClientFactory = new RelatednessClientFactory(vectorSpaceFactory);
+    public MongoIndraDriver(Params params, String mongoURI) {
+        super(params, new MongoVectorSpaceFactory(mongoURI), new MongoTranslator(mongoURI));
     }
 
-    public RelatednessResult getRelatedness(List<TextPair> pairs) {
-        return getRelatedness(pairs, currentParams);
-    }
-
-    public RelatednessResult getRelatedness(List<TextPair> pairs, Params params) {
-        RelatednessClient relatednessClient = relatednessClientFactory.create(params);
-        RelatednessResult result = relatednessClient.getRelatedness(pairs);
-
-        return result;
-    }
-
-    public Map<String, RealVector> getVectors(List<String> terms) {
-        return this.getVectors(terms, this.currentParams);
-    }
-
-    public Map<String, RealVector> getVectors(List<String> terms, Params params) {
-        VectorSpace vectorSpace = vectorSpaceFactory.create(params);
-        IndraAnalyzer analyzer = new IndraAnalyzer(params.language, false);
-
-        List<MutableAnalyzedTerm> analyzedTerms = new LinkedList<>();
-
-        for (String term : terms) {
-            try {
-                List<String> analyzedTokens = analyzer.analyze(term);
-                //analyzedTerms.add(new MutableAnalyzedTerm(term, analyzedTokens));
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-        }
-
-        Map<MutableAnalyzedTerm, Map<Integer, Double>> inVectors = vectorSpace.getVectors(analyzedTerms);
-        Map<String, RealVector> outVectors = new HashMap();
-
-        for (MutableAnalyzedTerm term : inVectors.keySet()) {
-            double[] dv = inVectors.get(term).values().stream().mapToDouble(d -> d).toArray();
-            if (vectorSpace.isSparse()) {
-                outVectors.put(term.getTerm(), new OpenMapRealVector(dv));
-            } else {
-                outVectors.put(term.getTerm(), new ArrayRealVector(dv));
-            }
-        }
-
-        return outVectors;
-    }
 }
